Stalled progress
It's taking too much time to figure out the right startup sequence for xfce4 - decided to upgrade barney's host operating system.

might pick this up again to provide  a simpler ros1 focal/noetic container for validation.

if so, recommend to myself to start from scratch  and switch to lxde with a single virtual display.